% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string}
\usage{
strsplit1(x, split = ",")
}
\arguments{
\item{x}{A single character vector}

\item{split}{Character to remove with split}
}
\value{
Vector of strings
}
\description{
Split a string
}
\examples{
 x <- "alfa,bravo,charlie,delta"
 output<-strsplit1(x,split=",")
 output[3]=="charlie"
 y<-"alfa,bravo,charlie,delta"
 strsplit1(y)
 strsplit1(y,split="a")
}
